Revert "Isort linting"
This reverts commit fbc1da762e0610ce4a6c49f3b295e8227de82038.
