+++ cross-reference/crossreference/cr/admin.py
@@ -1,5 +1,5 @@
-from cr.models import TestFailure, TestRun
 from django.contrib import admin
+from cr.models import TestFailure, TestRun
 
 admin.site.register(TestFailure)
 admin.site.register(TestRun)
