+++ cross-reference/crossreference/cr/views.py
@@ -1,6 +1,7 @@
+from django.shortcuts import render
+
 # Create your views here.
 from django.http import HttpResponse
-from django.shortcuts import render
 
 from .models import select_test_failures
