+++ cross-reference/crossreference/crossreference/settings.py
@@ -1,5 +1,4 @@
 import os
-
 from dotenv import load_dotenv
 
 # Load environment variables from .env file
