+++ define_masters.py
@@ -2,7 +2,6 @@
 
 import os
 import shutil
-
 import yaml
 
 BASE_PATH = "autogen/"
