+++ locks.py
@@ -1,12 +1,11 @@
+from buildbot.plugins import *
+from buildbot.process.properties import Property, Properties
+from buildbot.steps.shell import ShellCommand, Compile, Test, SetPropertyFromCommand
+from buildbot.steps.mtrlogobserver import MTR, MtrLogObserver
+from buildbot.steps.source.github import GitHub
+from buildbot.process.remotecommand import RemoteCommand
 from twisted.internet import defer
 
-from buildbot.plugins import *
-from buildbot.process.properties import Properties, Property
-from buildbot.process.remotecommand import RemoteCommand
-from buildbot.steps.mtrlogobserver import MTR, MtrLogObserver
-from buildbot.steps.shell import (Compile, SetPropertyFromCommand,
-                                  ShellCommand, Test)
-from buildbot.steps.source.github import GitHub
 from constants import *
 
 ####### LOCKS
