+++ master-web/buildbot.tac
@@ -1,7 +1,6 @@
 import os
 
 from twisted.application import service
-
 from buildbot.master import BuildMaster
 
 basedir = "."
@@ -23,8 +22,8 @@ if basedir == ".":
 # note: this line is matched against to check that this is a buildmaster
 # directory; do not edit it.
 application = service.Application("buildmaster")
-from twisted.python.log import FileLogObserver, ILogObserver
 from twisted.python.logfile import LogFile
+from twisted.python.log import ILogObserver, FileLogObserver
 
 logfile = LogFile.fromFullPath(
     os.path.join(log_basedir, "master-web.log"),
